requirements1.txt
asgiref==3.7.2
attrs==23.1.0
Django==4.2.5
django-cors-headers==4.2.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
inflection==0.5.1
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
mysqlclient==2.2.0
Pillow==10.0.1
PyJWT==2.8.0
pyotp==2.9.0
pypng==0.20220715.0
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
qrcode==7.4.2
referencing==0.30.
rpds-py==0.10.3
sqlparse==0.4.4
typing_extensions==4.8.0
uritemplate==4.1.1
